fix: refactor components folder structure gf-79
