+++ apps/frontend/src/libs/components/table/components/components.ts
@@ -1,0 +1,1 @@
+export { SelectRowCell } from "./select-row-cell/select-row-cell.js";
