+++ apps/frontend/src/pages/access-management/libs/components/components.ts
@@ -1,0 +1,4 @@
+export { GroupCreateForm } from "./group-create-form/group-create-form.js";
+export { GroupUsersTable } from "./group-users-table/group-users-table.js";
+export { GroupsTable } from "./groups-table/groups-table.js";
+export { UsersTable } from "./users-table/users-table.js";
