+++ apps/frontend/src/pages/access-management/libs/components/group-create-form/libs/helpers/get-permission-options.helper.ts
