+++ apps/frontend/src/pages/access-management/libs/components/group-create-form/libs/helpers/helpers.ts
