+++ apps/frontend/src/pages/access-management/libs/components/groups-table/groups-table.tsx
@@ -1,35 +1,35 @@
 import { Table, TablePagination } from "~/libs/components/components.js";
-import { type UserGetAllItemResponseDto } from "~/modules/users/users.js";
+import { type GroupGetAllItemResponseDto } from "~/modules/groups/groups.js";
 
-import { getUserColumns, getUserRows } from "../../libs/helpers/helpers.js";
-import { type UserRow } from "../../libs/types/types.js";
+import { getGroupColumns, getGroupRows } from "../../helpers/helpers.js";
+import { type GroupRow } from "../../types/types.js";
 import styles from "./styles.module.css";
 
 type Properties = {
+	groups: GroupGetAllItemResponseDto[];
 	onPageChange: (page: number) => void;
 	onPageSizeChange: (pageSize: number) => void;
 	page: number;
 	pageSize: number;
 	paginationBackground?: "primary" | "secondary";
 	totalItemsCount: number;
-	users: UserGetAllItemResponseDto[];
 };
 
-const UsersTable = ({
+const GroupsTable = ({
+	groups,
 	onPageChange,
 	onPageSizeChange,
 	page,
 	pageSize,
 	paginationBackground = "primary",
 	totalItemsCount,
-	users,
 }: Properties): JSX.Element => {
-	const userColumns = getUserColumns();
-	const userData: UserRow[] = getUserRows(users);
+	const groupColumns = getGroupColumns();
+	const groupData: GroupRow[] = getGroupRows(groups);
 
 	return (
-		<div className={styles["users-table"]}>
-			<Table<UserRow> columns={userColumns} data={userData} />
+		<div className={styles["groups-table"]}>
+			<Table<GroupRow> columns={groupColumns} data={groupData} />
 			<TablePagination
 				background={paginationBackground}
 				onPageChange={onPageChange}
@@ -42,4 +42,4 @@ const UsersTable = ({
 	);
 };
 
-export { UsersTable };
+export { GroupsTable };
