+++ apps/frontend/src/pages/access-management/libs/components/groups-table/styles.module.css
