+++ apps/frontend/src/pages/access-management/libs/components/users-table/users-table.tsx
@@ -1,35 +1,35 @@
 import { Table, TablePagination } from "~/libs/components/components.js";
-import { type GroupGetAllItemResponseDto } from "~/modules/groups/groups.js";
+import { type UserGetAllItemResponseDto } from "~/modules/users/users.js";
 
-import { getGroupColumns, getGroupRows } from "../../libs/helpers/helpers.js";
-import { type GroupRow } from "../../libs/types/types.js";
+import { getUserColumns, getUserRows } from "../../helpers/helpers.js";
+import { type UserRow } from "../../types/types.js";
 import styles from "./styles.module.css";
 
 type Properties = {
-	groups: GroupGetAllItemResponseDto[];
 	onPageChange: (page: number) => void;
 	onPageSizeChange: (pageSize: number) => void;
 	page: number;
 	pageSize: number;
 	paginationBackground?: "primary" | "secondary";
 	totalItemsCount: number;
+	users: UserGetAllItemResponseDto[];
 };
 
-const GroupsTable = ({
-	groups,
+const UsersTable = ({
 	onPageChange,
 	onPageSizeChange,
 	page,
 	pageSize,
 	paginationBackground = "primary",
 	totalItemsCount,
+	users,
 }: Properties): JSX.Element => {
-	const groupColumns = getGroupColumns();
-	const groupData: GroupRow[] = getGroupRows(groups);
+	const userColumns = getUserColumns();
+	const userData: UserRow[] = getUserRows(users);
 
 	return (
-		<div className={styles["groups-table"]}>
-			<Table<GroupRow> columns={groupColumns} data={groupData} />
+		<div className={styles["users-table"]}>
+			<Table<UserRow> columns={userColumns} data={userData} />
 			<TablePagination
 				background={paginationBackground}
 				onPageChange={onPageChange}
@@ -42,4 +42,4 @@ const GroupsTable = ({
 	);
 };
 
-export { GroupsTable };
+export { UsersTable };
